chore: make _version_ columns integers, #128

+++ packages/database-postgres/entity/Account.ts
@@ -65,7 +65,7 @@ export class Account {
 
   @Column({
     default: 0,
-    type: 'bigint',
+    type: 'integer',
     nullable: false,
   })
   public _version_: number;
