+++ packages/database-postgres/entity/Issuer.ts
@@ -35,7 +35,7 @@ export class Issuer {
 
   @Column({
     default: 0,
-    type: 'bigint',
+    type: 'integer',
     nullable: false,
   })
   public _version_: number;
